requirements.txt
# Production Quantum Computing Platform Dependencies
# Cirq + TensorFlow based quantum computing stack

# Core Quantum Computing
cirq>=1.6.0
tensorflow>=2.20.0
tensorflow-quantum>=0.7.3
sympy>=1.14.0

# Scientific Computing
numpy>=1.26.0
scipy>=1.16.0
matplotlib>=3.8.0

# Machine Learning & AI
scikit-learn>=1.3.0
quantum-ml>=0.1.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Network Simulation & Graph Theory
networkx>=3.1
simpy>=4.0.1
igraph>=0.10.0
graph-tool>=2.45

# Cryptography & Security
cryptography>=41.0.0
pycryptodome>=3.18.0
pynacl>=1.5.0
cryptg>=0.4.0
bcrypt>=4.0.0

# Web Framework & API
fastapi>=0.103.0
uvicorn>=0.23.0
websockets>=11.0.0
jinja2>=3.1.0
httpx>=0.24.0

# Data Processing & Analytics
pandas>=2.1.0
polars>=0.19.0
dask>=2023.8.0
joblib>=1.3.0

# Visualization & UI
plotly>=5.15.0
bokeh>=3.2.0
dash>=2.14.0
streamlit>=1.26.0
gradio>=3.44.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=4.6.0
pymongo>=4.5.0

# Configuration & Environment
pydantic>=2.3.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.30.0

# Async & Concurrency
asyncio>=3.4.3
aiohttp>=3.8.0
aiofiles>=23.2.0
celery>=5.3.0

# CLI & Utilities
click>=8.1.0
typer>=0.9.0
rich>=13.5.0
tqdm>=4.66.0
psutil>=5.9.0

# Testing & Quality
hypothesis>=6.82.0
factory-boy>=3.3.0
faker>=19.6.0
coverage>=7.3.0

# Documentation
sphinx>=7.1.0
mkdocs>=1.5.0
mkdocs-material>=9.2.0